refactor: 공통 hook 폴더 분리
